dockerized the service locally, optmized by adding redis caching, optmized the code style and updated readme
