+++ .env.example
@@ -1,6 +1,10 @@
-GOOGLE_API_KEY = ""
-DB_USERNAME = ""
-DB_PASSWORD = ""
-DB_HOST = ""
-DB_PORT = ""
-DB_NAME = ""
+GOOGLE_API_KEY=""
+DB_USERNAME=""
+DB_PASSWORD=""
+DB_HOST=""
+DB_PORT=""
+DB_NAME=""
+GOOGLE_lOCATIONS_APP_PORT=""
+REDIS_HOST=""
+REDIS_HOST_PORT=""
+DOCKER_NETWORK=""
