+++ README.md
@@ -1,15 +1,15 @@
 # google_locations
 Project that retrieves google locations of GPs
-Instructions on how to run the script:
-1. Setup the "platform" database locally by the running the migrations for it: https://github.com/Givelify/platform-db-migrator
-2. Copy the contents of .env.example file and add database credentials and google api key to run this project
-3. Setup a virtual environment
-4. Run 'pip install -r requirements.txt' in terminal
-5. Run the main script using "python main.py" if you want to process multiple GPs in the donee_info table or 
- - "python main.py --id {ID}" if you want to process a specific GP from donee_info table by providing their ID.
- - If you want autocomplete check enabled use the additional parameter "--enable_autocomplete" in the command. Autocomplete check is disabled by default.
- Examples of ways to run:
- 'python main.py', processes multiple GPs at once
- 'python main.py --id 479", processes GP with ID 479
- 'python main.py --id 479 --enable_autocomplete', processes GP with ID 479 while utilizing the autocomplete API
- 'python main.py --enable_autocomplete' processes multiple GPs at once while utilizing the autocomplete API
+
+Spin this project up in docker through the following steps:
+
+ 1. setup environment variables: "cp example.env .env" and update the values
+ 2. Make sure your MySQL container is already running and connected to the same Docker network as "DOCKER_NETWORK" in the .env file if you're using an external database container
+ 3. build and run the services using by running the "docker compose up --build -d" command in the path/to/google-locations directory
+ 4. Then run the script using "docker compose exec google-locations-app python main.py {optional arguments}"
+
+ The optional command line args you can use are:
+ 1. --id <ID>: If you want to process a specific GP from donee_info table by providing their ID
+ 2. --enable_autocomplete: processes GPs while utilizing autocomplete API. Autocomplete check is disabled by default.
+ 3. --cache_check <Bool>: Enables Redis cache check to see if a GP has previosuly failed to avoid running them again. The redis cache stores the GP IDs for a month from when they first failed. Use '--cache_check IDs cache_check is True by default
+
