+++ config.py
@@ -26,3 +26,6 @@ class Config:
     logger = slogger.StructuredLogger.getLogger(
         "google-locations", "INFO", stdout_handler
     )
+    REDIS_HOST=os.getenv("REDIS_HOST")
+    REDIS_HOST_PORT=os.getenv("REDIS_HOST_PORT")
+    REDIS_DB = os.getenv("REDIS_DB")
